requirements.txt
flask
spotipy
Pillow